moved the http client to more accessible for integrations
