+++ changelog/improvement.md
@@ -1,0 +1,1 @@
+Moved the ocean httpx client so it will be more accessible to integrations
