+++ port_ocean/clients/port/mixins/blueprints.py
@@ -4,7 +4,7 @@ import httpx
 from loguru import logger
 
 from port_ocean.clients.port.authentication import PortAuthentication
-from port_ocean.clients.port.utils import handle_status_code
+from port_ocean.clients.utils import handle_status_code
 from port_ocean.core.models import Blueprint
