+++ port_ocean/clients/port/mixins/entities.py
@@ -5,7 +5,7 @@ from loguru import logger
 
 from port_ocean.clients.port.authentication import PortAuthentication
 from port_ocean.clients.port.types import RequestOptions, UserAgentType
-from port_ocean.clients.port.utils import handle_status_code
+from port_ocean.clients.utils import handle_status_code
 from port_ocean.core.models import Entity
